Another fruitful commit. Even include documentation!
Add: ability to store a layout; ability to read layout from the file; ability to control the number of (extra/timed) queries to run; ability to use a single algorithm and whether or not to process the queries (for external profilers); new implementations on vEB that uses BFS index conversion; documentation.

Fix: testing; bugs in DFS searching algorithm; memory management.

Refactor: renaming for consistent naming convention with C++ standard libraries; reorganization of modules; .